👩‍💻 Deprecate `codeBlockPlugin` for captions
