+++ .changeset/warm-teachers-yawn.md
@@ -1,0 +1,7 @@
+---
+'myst-cli': patch
+'myst-directives': patch
+'myst-transforms': patch
+---
+
+Deprecate codeBlockPlugin for the caption parsing, which now happens in myst-parser and myst-directives
