+++ packages/myst-transforms/src/basic.ts
@@ -7,14 +7,11 @@ import { captionParagraphTransform } from './caption';
 import { admonitionHeadersTransform } from './admonitions';
 import { blockMetadataTransform, blockNestingTransform } from './blocks';
 import { htmlIdsTransform } from './htmlIds';
-import { codeBlockTransform } from './codeBlock';
 import { imageAltTextTransform } from './images';
 import { mathLabelTransform, mathNestingTransform } from './math';
 import { blockquoteTransform } from './blockquote';
 
 export function basicTransformations(tree: Root, file: VFile) {
-  // Must happen first
-  codeBlockTransform(tree); // TODO: ideally move this to the parser
   // lifting roles and directives must happen before the mystTarget transformation
   liftMystDirectivesAndRolesTransform(tree);
   // Some specifics about the ordering are noted below
